Initial version.
Code test:
"Write a Modern / Metro app in C#/XAML that consumes the feed (https://api.flickr.com/services/feeds/photos_public.gne?format=json&id=67710869@N08) and for each entry in the array items take the media.m (which is an address to a photo) and show it on the screen.
It’s up to the candidate how to show the photos: List? Grid? No preference."